More predictable output shape of filter function
`filter_outliers` filters the exact number of outliers given if possible
even if the maximum search rounds are reached without finding the
perfect neighbor threshold. If the number of ouliers is to low, another
search round were it was above is restored. Then, if the number is above the
desired number of outliers, the most distant records are removed until
the expected number of outliers is reached.